+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.9.0
 
 - **New features**
-  - Reset multiple users' passord button on User Settings
+  - Reset multiple users' password button on User Settings
 
 ## 1.8.3